Update EXP3_7229_Raghav_D2.cpp
Create a class Room with Data members: length, breadth. Include functions get () and display () to get and display the data of room. Include two additional functions: area () and perimeter () to calculate the area and perimeter of the room. Implement the class by creating at least 5 objects of the class and display the area
and perimeter of all 5 rooms.